pyproject.toml
[tool.poetry]
name = "django-orm-evalute"
version = "0.1.0"
description = ""
authors = ["문정훈 <black2st11@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.4"
django-debug-toolbar = "^3.4.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
